Set up proxy to relay api requests
