+++ assets/src/state/actions/auth.js
@@ -1,9 +1,10 @@
-import { createAction } from "redux-actions";
+import { createAction } from "redux-actions"
 
-import { CHANGE_COLOUR } from "./types";
+import { CHANGE_COLOUR } from "./types"
 
-const change_colour = createAction(CHANGE_COLOUR);
+const change_colour = createAction(CHANGE_COLOUR)
 
+// eslint-disable-next-line
 export const changeColour = () => dispatch => {
-  dispatch(change_colour());
-};
+  dispatch(change_colour())
+}
