+++ lib/your_sanctuary_web/controllers/page_controller.ex
@@ -2,6 +2,7 @@ defmodule YourSanctuaryWeb.PageController do
   use YourSanctuaryWeb, :controller
 
   def index(conn, _params) do
-    render conn, "index.html"
+    conn
+    |> json(%{great: "success"})
   end
 end
